+++ jb-core/src/main/resources/core-jbmessages_en.properties
@@ -125,3 +125,39 @@ discord.plurals.second[one] = second
 discord.plurals.second[other] = seconds
 
 discord.bonus.feature = This bonus feature is only available for servers whose owners supports us! :yellow_heart:\nWant to be one of them? [Check this out!]({0}/donate)
+
+welcome.guild.message = Hello! <:paws:403495496137703434>\n\nThank you for adding me to your server **{0}**! <:foxy:403496944485400577> \n\n\
+You can configure me using fancy dashboard at https://juniperbot.ru\n\
+Here is short review that I can do:
+#-
+welcome.fields.common.title = :gear: Common
+welcome.fields.common.content = You can change prefix, command and interface language, configure some moderation behaviour. English and Russian languages are supported for both dashboard and bot itself!
+#-
+welcome.fields.music.title = :gear: Music
+welcome.fields.music.content = You can play music from most common sources like YouTube, SoundCloud, BandCamp, Vimeo, Twitch.tv and even some radiostations. You can configure behaviour like restrict music controlling to specified DJ-Roles, configure playback limitations and more!
+#-
+welcome.fields.audit.title = :books: Audit
+welcome.fields.audit.content = You can keep history of member actions like moderation, message delete/edit, member joining/leaving and more. You can review them in dashboard or even in separate logging channel on your server!
+#-
+welcome.fields.subscriptions.title = :loudspeaker: Subscriptions
+welcome.fields.subscriptions.content = You can configure subscriptions to your own Vkontakte community or any Twitch.tv and YouTube channels. Here is JuniperFoxx Instagram subscription too. Yip!
+#-
+welcome.fields.ranking.title = :trophy: Member Ranking
+welcome.fields.ranking.content = You can give your members some rewards for their activity on the server!
+#-
+welcome.fields.welcome.title = :wave: Welcome messages
+welcome.fields.welcome.content = You can configure member welcome messages, assign starting roles, restore old roles and nickname if member returns after leaving and more.
+#-
+welcome.fields.commands.title = :speech_balloon: Command
+welcome.fields.commands.content = You can configure each command behaviour - cooldown, roles and channels for which commands would work
+#-
+welcome.fields.custom.title = :pencil: Custom commands
+welcome.fields.custom.content = You can even add your own simple commands with message response or as an alias to other command. They are configurable too just like internal commands
+#-
+welcome.fields.fun.title = :performing_arts: Fun
+welcome.fields.fun.content = There are also some small fun features like reaction roulette, cookie karma, dogs/cats pictures, 8ball and more
+#-
+welcome.fields.support.title = :question: Have a question?
+welcome.fields.support.content = **Our website:** {0}\n\
+**Our Discord official server:** {1}\n\
+**Our GitHub page:** {2}
