+++ jb-core/src/main/resources/core-jbmessages_ru.properties
@@ -133,3 +133,39 @@ discord.plurals.second[many] = секунд
 discord.plurals.second[other] = секунд
 
 discord.bonus.feature = Эта бонусная функция доступна только для серверов, чьи владельцы поддерживают нас! :yellow_heart:\nХотите быть одними из них? [Поддержите нас!]({0}/patreon)
+
+welcome.guild.message = Привет! <:paws:403495496137703434>\n\nСпасибо, что добавили меня на свой сервер **{0}**! <:foxy:403496944485400577> \n\n\
+Вы можете настроить меня в удобной панели управления здесь: https://juniperbot.ru\n\
+Вот краткий список того, что я умею:
+#-
+welcome.fields.common.title = :gear: Основное
+welcome.fields.common.content = Вы можете поменять префикс, язык команд и интерфейса, настроить модерирование. Мной поддерживаются Английский и Русский языки как на сайте, так и здесь!
+#-
+welcome.fields.music.title = :musical_note: Музыка
+welcome.fields.music.content = Слушайте музыку из основных источников, таких как YouTube, SoundCloud, BandCamp, Vimeo, Twitch.tv и даже некоторые радиостанции. Вы можете настроить музыку как вам угодно: ограничить управление DJ-ролями, настроить ограничения на треки и ещё много всякого!
+#-
+welcome.fields.audit.title = :books: Аудит
+welcome.fields.audit.content = Храните историю действий участиков, таких как удаление/редактирование сообщений, заходы/выходы участиков и прочее. Вы можете просмотреть хронологию на сайте или настроить отправку логов в канал сервера!
+#-
+welcome.fields.subscriptions.title = :loudspeaker: Подписки
+welcome.fields.subscriptions.content = Подпишитесь на ваши сообщества Вконтакте или любимые каналы Twitch.tv и YouTube. Ну и, конечно же, куда без замечательной лисички JuniperFoxx. Йип!
+#-
+welcome.fields.ranking.title = :trophy: Member Рейтинг участников
+welcome.fields.ranking.content = Поощряйте участников за их активное участие на сервере выдачей ролей за уровни и прочее!
+#-
+welcome.fields.welcome.title = :wave: Приветствия
+welcome.fields.welcome.content = Настройте приветствие новых пользователей, выдачу им начальных ролей, восстановление старых при перезаходе и многое другое.
+#-
+welcome.fields.commands.title = :speech_balloon: Команды
+welcome.fields.commands.content = Настройте поведение любых команд: их кулдаун, доступные для использования роли и каналы.
+#-
+welcome.fields.custom.title = :pencil: Пользовательские команды
+welcome.fields.custom.content = Добавьте свои собственные команды с ответным сообщением или перенаправлением на другую. У них есть все аналогичные настройки, как и у встроенных команд.
+#-
+welcome.fields.fun.title = :performing_arts: Весёлости
+welcome.fields.fun.content = Здесь также есть некоторые полезные, бесполезные и весёлые мелочи вроде смайл-рулетки, печеньки-карма, фотки котиков, собачек, магический шар и многое другое!
+#-
+welcome.fields.support.title = :question: Возникли вопросы?
+welcome.fields.support.content = **Наш сайт:** {0}\n\
+**Наш официальный сервер в Discord:** {1}\n\
+**Наша страничка на GitHub:** {2}
