--- modules/jb-module-welcome/src/main/resources/welcome-jbmessages_en.properties
@@ -1,22 +1,0 @@
-welcome.guild.message = Hello `{0}`! <:paws:403495496137703434>\n\nThank you for adding me to your server `{1}`! <:foxy:403496944485400577> \n\n\
-Please visit https://juniperbot.ru to configure all available features:
-#-
-welcome.fields.common.title = :gear: Common configuration:
-welcome.fields.common.content = In this section you can set prefix, language, music and publishing options\n**English and Russian languages are supported**
-#-
-welcome.fields.ranking.title = :trophy: Ranking configuration:
-welcome.fields.ranking.content = In this section you can configure ranking and level rewards
-#-
-welcome.fields.welcome.title = :wave: Welcome messages:
-welcome.fields.welcome.content = In this section you can configure member welcome and leave messages
-#-
-welcome.fields.commands.title = :speech_balloon: Command list:
-welcome.fields.commands.content = In this section you can configure all available commands just in one place
-#-
-welcome.fields.custom.title = :pencil: Custom commands:
-welcome.fields.custom.content = In this section you can add your own commands with some info or something else
-#-
-welcome.fields.support.title = :question: Have a question?
-welcome.fields.support.content = **Our website:** {0}\n\
-**Our Discord official server:** {1}\n\
-**Our GitHub page:** {2}
