--- modules/jb-module-welcome/src/main/resources/welcome-jbmessages_ru.properties
@@ -1,22 +1,0 @@
-welcome.guild.message = Привет `{0}`! <:paws:403495496137703434>\n\nСпасибо, что добавили меня на свой сервер `{1}`! \nТеперь тут будет чуть фыр-фырнее <:foxy:403496944485400577> \n\n\
-Чтобы настроить меня по Вашему вкусу и для Ваших нужд, посетите https://juniperbot.ru Там вы можете найти:
-#-
-welcome.fields.common.title = :gear: Основные настройки:
-welcome.fields.common.content = Настройте префикс и язык бота, музыкальные возможности и возможности публикации\n**Поддерживаются английский и русский языки**
-#-
-welcome.fields.ranking.title = :trophy: Рейтинг участников:
-welcome.fields.ranking.content = Настройте рейтинг активности участников, а также награды за активность
-#-
-welcome.fields.welcome.title = :wave: Приветствия:
-welcome.fields.welcome.content = Настройте сообщения с приветствием новых участников, а также покинувших сервер
-#-
-welcome.fields.commands.title = :speech_balloon: Команды:
-welcome.fields.commands.content = Настройте в одном месте все доступные команды на сервере
-#-
-welcome.fields.custom.title = :pencil: Пользовательские команды:
-welcome.fields.custom.content = Настройте свои собственные команды, которые либо выполнят другие команды, либо отобразят какое-либо сообщение
-#-
-welcome.fields.support.title = :question: Возникли вопросы?
-welcome.fields.support.content = **Наш сайт:** {0}\n\
-**Наш официальный сервер в Discord:** {1}\n\
-**Наша страничка на GitHub:** {2}
